Make WiseAgentsLoader set up YAML env var loading automatically
