+++ src/wiseagents/yaml/__init__.py
@@ -1,12 +1,12 @@
 # This is the __init__.py file for the wiseagents.yaml package
 
 # Import any modules or subpackages here
-from .yaml_utils import setup_yaml_for_env_vars
-from .wise_yaml_loader import WiseAgentsLoader
+from .wise_yaml_loader import WiseAgentsLoader, setup_yaml_for_env_vars
 
 
 # Define any necessary initialization code here
-
+setup_yaml_for_env_vars()
 # Optionally, you can define __all__ to specify the public interface of the package
 # __all__ = ['module1', 'module2', 'subpackage']
-__all__ = ['setup_yaml_for_env_vars', 'WiseAgentsLoader']
+__all__ = ['WiseAgentsLoader']
+
